Fix post rendering
When using HEEx + a post (meaning, some HTML generated from markdown),
it attempts to render plain HTML, which HEEx protects you from. You
can't just pipe it all through "Phoenix.HTML.raw", as that seems to
crash on content that is HEEx and not a string.

Solution: create a component that branches on whether the content is a
string or not, and respond accordingly.